Seems to work, so replacing it.
