+++ mylib/truecaser.py
@@ -1,41 +1,21 @@
 from typing import Dict, Optional, List, Any
 
-import numpy
 from overrides import overrides
-import torch
-from torch.nn.modules.linear import Linear
-import torch.nn.functional as F
 
-from allennlp.common.checks import check_dimensions_match
 from allennlp.data import Vocabulary
-from allennlp.modules import Seq2SeqEncoder, TimeDistributed, TextFieldEmbedder
+from allennlp.modules import Seq2SeqEncoder, TextFieldEmbedder
 from allennlp.models.model import Model
+from allennlp.models.simple_tagger import SimpleTagger
 from allennlp.nn import InitializerApplicator, RegularizerApplicator
-from allennlp.nn.util import get_text_field_mask, sequence_cross_entropy_with_logits
-from allennlp.training.metrics import CategoricalAccuracy, F1Measure
+from allennlp.training.metrics import F1Measure
 
 
-@Model.register("truecaser-old")
-class TrueCaser(Model):
+@Model.register("truecaser")
+class TrueCaser(SimpleTagger):
     """
-    Basically just the simple tagger but with f1 measure.
-
-    This ``SimpleTagger`` simply encodes a sequence of text with a stacked ``Seq2SeqEncoder``, then
-    predicts a tag for each token in the sequence.
-
-    Parameters
-    ----------
-    vocab : ``Vocabulary``, required
-        A Vocabulary, required in order to compute sizes for input/output projections.
-    text_field_embedder : ``TextFieldEmbedder``, required
-        Used to embed the ``tokens`` ``TextField`` we get as input to the model.
-    encoder : ``Seq2SeqEncoder``
-        The encoder (with its own internal stacking) that we will use in between embedding tokens
-        and predicting output tags.
-    initializer : ``InitializerApplicator``, optional (default=``InitializerApplicator()``)
-        Used to initialize the model parameters.
-    regularizer : ``RegularizerApplicator``, optional (default=``None``)
-        If provided, will be used to calculate the regularization penalty during training.
+    This is a simple child of SimpleTagger. The only difference is that
+    I wanted to include an F1 measure in the metrics (even though this is
+    character F1 and not token F1).
     """
 
     def __init__(self, vocab: Vocabulary,
@@ -43,104 +23,8 @@ class TrueCaser(Model):
                  encoder: Seq2SeqEncoder,
                  initializer: InitializerApplicator = InitializerApplicator(),
                  regularizer: Optional[RegularizerApplicator] = None) -> None:
-        super(TrueCaser, self).__init__(vocab, regularizer)
-
-        self.text_field_embedder = text_field_embedder
-        self.num_classes = self.vocab.get_vocab_size("labels")
-        self.encoder = encoder
-        self.tag_projection_layer = TimeDistributed(Linear(self.encoder.get_output_dim(),
-                                                           self.num_classes))
-
-        check_dimensions_match(text_field_embedder.get_output_dim(), encoder.get_input_dim(),
-                               "text field embedding dim", "encoder input dim")
-        self.metrics = {
-                "accuracy": CategoricalAccuracy(),
-                "accuracy3": CategoricalAccuracy(top_k=3),
-                "f1": F1Measure(positive_label=self.vocab.get_token_index("U", namespace="labels"))
-        }
-
-        initializer(self)
-
-    @overrides
-    def forward(self,  # type: ignore
-                tokens: Dict[str, torch.LongTensor],
-                tags: torch.LongTensor = None,
-                metadata: List[Dict[str, Any]] = None) -> Dict[str, torch.Tensor]:
-        # pylint: disable=arguments-differ
-        """
-        Parameters
-        ----------
-        tokens : Dict[str, torch.LongTensor], required
-            The output of ``TextField.as_array()``, which should typically be passed directly to a
-            ``TextFieldEmbedder``. This output is a dictionary mapping keys to ``TokenIndexer``
-            tensors.  At its most basic, using a ``SingleIdTokenIndexer`` this is: ``{"tokens":
-            Tensor(batch_size, num_tokens)}``. This dictionary will have the same keys as were used
-            for the ``TokenIndexers`` when you created the ``TextField`` representing your
-            sequence.  The dictionary is designed to be passed directly to a ``TextFieldEmbedder``,
-            which knows how to combine different word representations into a single vector per
-            token in your input.
-        tags : torch.LongTensor, optional (default = None)
-            A torch tensor representing the sequence of integer gold class labels of shape
-            ``(batch_size, num_tokens)``.
-        metadata : ``List[Dict[str, Any]]``, optional, (default = None)
-            metadata containing the original words in the sentence to be tagged under a 'words' key.
-
-        Returns
-        -------
-        An output dictionary consisting of:
-        logits : torch.FloatTensor
-            A tensor of shape ``(batch_size, num_tokens, tag_vocab_size)`` representing
-            unnormalised log probabilities of the tag classes.
-        class_probabilities : torch.FloatTensor
-            A tensor of shape ``(batch_size, num_tokens, tag_vocab_size)`` representing
-            a distribution of the tag classes per word.
-        loss : torch.FloatTensor, optional
-            A scalar loss to be optimised.
-
-        """
-        embedded_text_input = self.text_field_embedder(tokens)
-        batch_size, sequence_length, _ = embedded_text_input.size()
-        mask = get_text_field_mask(tokens)
-        encoded_text = self.encoder(embedded_text_input, mask)
-
-        logits = self.tag_projection_layer(encoded_text)
-        reshaped_log_probs = logits.view(-1, self.num_classes)
-        class_probabilities = F.softmax(reshaped_log_probs, dim=-1).view([batch_size,
-                                                                          sequence_length,
-                                                                          self.num_classes])
-
-        output_dict = {"logits": logits, "class_probabilities": class_probabilities}
-
-        if tags is not None:
-            loss = sequence_cross_entropy_with_logits(logits, tags, mask)
-            for metric in self.metrics.values():
-                metric(logits, tags, mask.float())
-            output_dict["loss"] = loss
-
-        if metadata is not None:
-            output_dict["words"] = [x["words"] for x in metadata]
-        return output_dict
-
-    @overrides
-    def decode(self, output_dict: Dict[str, torch.Tensor]) -> Dict[str, torch.Tensor]:
-        """
-        Does a simple position-wise argmax over each token, converts indices to string labels, and
-        adds a ``"tags"`` key to the dictionary with the result.
-        """
-        all_predictions = output_dict['class_probabilities']
-        all_predictions = all_predictions.cpu().data.numpy()
-        if all_predictions.ndim == 3:
-            predictions_list = [all_predictions[i] for i in range(all_predictions.shape[0])]
-        else:
-            predictions_list = [all_predictions]
-        all_tags = []
-        for predictions in predictions_list:
-            argmax_indices = numpy.argmax(predictions, axis=-1)
-            tags = [self.vocab.get_token_from_index(x, namespace="labels")
-                    for x in argmax_indices]
-            all_tags.append(tags)
-        output_dict['tags'] = all_tags
-        return output_dict
+        super(TrueCaser, self).__init__(vocab, text_field_embedder, encoder, initializer, regularizer)
+        self.metrics["f1"] = F1Measure(positive_label=self.vocab.get_token_index("U", namespace="labels"))
 
     @overrides
     def get_metrics(self, reset: bool = False) -> Dict[str, float]:
